+++ projects/_02_sfc_internal/unittest_data/sfc_test_file_schemas.py
@@ -3,6 +3,7 @@ from dataclasses import dataclass
 from pyspark.sql.types import (
     ArrayType,
     DateType,
+    FloatType,
     IntegerType,
     StringType,
     StructField,
@@ -1207,5 +1208,68 @@ class FlattenCQCRatings:
         [
             StructField(CQCL.location_id, StringType(), True),
             StructField(CQCRatings.location_id_hash, StringType(), True),
+        ]
+    )
+
+
+@dataclass
+class LmEngagementUtilsSchemas:
+    add_columns_for_locality_manager_dashboard_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField(IndCQC.current_cssr, StringType(), True),
+            StructField(CoverageColumns.in_ascwds, IntegerType(), True),
+            StructField(Keys.year, StringType(), True),
+        ]
+    )
+    expected_add_columns_for_locality_manager_dashboard_schema = StructType(
+        [
+            *add_columns_for_locality_manager_dashboard_schema,
+            StructField(CoverageColumns.la_monthly_coverage, FloatType(), True),
+            StructField(CoverageColumns.coverage_monthly_change, FloatType(), True),
+            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
+            StructField(CoverageColumns.new_registrations_monthly, IntegerType(), True),
+            StructField(CoverageColumns.new_registrations_ytd, IntegerType(), True),
+        ]
+    )
+
+    expected_calculate_la_coverage_monthly_schema = StructType(
+        [
+            *add_columns_for_locality_manager_dashboard_schema,
+            StructField(CoverageColumns.la_monthly_coverage, FloatType(), True),
+        ]
+    )
+    calculate_coverage_monthly_change_schema = (
+        expected_calculate_la_coverage_monthly_schema
+    )
+
+    expected_calculate_coverage_monthly_change_schema = StructType(
+        [
+            *expected_calculate_la_coverage_monthly_schema,
+            StructField(CoverageColumns.coverage_monthly_change, FloatType(), True),
+        ]
+    )
+
+    calculate_locations_monthly_change_schema = (
+        expected_calculate_coverage_monthly_change_schema
+    )
+    expected_calculate_locations_monthly_change_schema = StructType(
+        [
+            *expected_calculate_coverage_monthly_change_schema,
+            StructField(CoverageColumns.in_ascwds_last_month, IntegerType(), True),
+            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
+        ]
+    )
+
+    calculate_new_registrations_schema = (
+        expected_calculate_locations_monthly_change_schema
+    )
+    expected_calculate_new_registrations_schema = StructType(
+        [
+            *expected_calculate_coverage_monthly_change_schema,
+            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
+            StructField(CoverageColumns.new_registrations_monthly, IntegerType(), True),
+            StructField(CoverageColumns.new_registrations_ytd, IntegerType(), True),
         ]
     )
