+++ tests/test_file_schemas.py
@@ -467,69 +467,6 @@ class FilterCleanedValuesSchema:
 
 
 @dataclass
-class LmEngagementUtilsSchemas:
-    add_columns_for_locality_manager_dashboard_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.current_cssr, StringType(), True),
-            StructField(CoverageColumns.in_ascwds, IntegerType(), True),
-            StructField(Keys.year, StringType(), True),
-        ]
-    )
-    expected_add_columns_for_locality_manager_dashboard_schema = StructType(
-        [
-            *add_columns_for_locality_manager_dashboard_schema,
-            StructField(CoverageColumns.la_monthly_coverage, FloatType(), True),
-            StructField(CoverageColumns.coverage_monthly_change, FloatType(), True),
-            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
-            StructField(CoverageColumns.new_registrations_monthly, IntegerType(), True),
-            StructField(CoverageColumns.new_registrations_ytd, IntegerType(), True),
-        ]
-    )
-
-    expected_calculate_la_coverage_monthly_schema = StructType(
-        [
-            *add_columns_for_locality_manager_dashboard_schema,
-            StructField(CoverageColumns.la_monthly_coverage, FloatType(), True),
-        ]
-    )
-    calculate_coverage_monthly_change_schema = (
-        expected_calculate_la_coverage_monthly_schema
-    )
-
-    expected_calculate_coverage_monthly_change_schema = StructType(
-        [
-            *expected_calculate_la_coverage_monthly_schema,
-            StructField(CoverageColumns.coverage_monthly_change, FloatType(), True),
-        ]
-    )
-
-    calculate_locations_monthly_change_schema = (
-        expected_calculate_coverage_monthly_change_schema
-    )
-    expected_calculate_locations_monthly_change_schema = StructType(
-        [
-            *expected_calculate_coverage_monthly_change_schema,
-            StructField(CoverageColumns.in_ascwds_last_month, IntegerType(), True),
-            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
-        ]
-    )
-
-    calculate_new_registrations_schema = (
-        expected_calculate_locations_monthly_change_schema
-    )
-    expected_calculate_new_registrations_schema = StructType(
-        [
-            *expected_calculate_coverage_monthly_change_schema,
-            StructField(CoverageColumns.locations_monthly_change, IntegerType(), True),
-            StructField(CoverageColumns.new_registrations_monthly, IntegerType(), True),
-            StructField(CoverageColumns.new_registrations_ytd, IntegerType(), True),
-        ]
-    )
-
-
-@dataclass
 class IndCQCDataUtils:
     input_schema_for_adding_estimate_filled_posts_and_source = StructType(
         [
